@@ -4,18 +4,18 @@
   "license": "Apache-2.0",
   "description": "Gravitee.io APIM - UI Console",
   "dependencies": {
-    "@angular/animations": "17.2.3",
+    "@angular/animations": "18.2.0",
     "@angular/cdk": "17.2.1",
-    "@angular/common": "17.2.3",
-    "@angular/compiler": "17.2.3",
-    "@angular/core": "17.2.3",
-    "@angular/forms": "17.2.3",
+    "@angular/common": "18.2.0",
+    "@angular/compiler": "18.2.0",
+    "@angular/core": "18.2.0",
+    "@angular/forms": "18.2.0",
     "@angular/material": "17.2.1",
     "@angular/material-moment-adapter": "17.2.1",
-    "@angular/platform-browser": "17.2.3",
-    "@angular/platform-browser-dynamic": "17.2.3",
-    "@angular/router": "17.2.3",
-    "@angular/upgrade": "17.2.3",
+    "@angular/platform-browser": "18.2.0",
+    "@angular/platform-browser-dynamic": "18.2.0",
+    "@angular/router": "18.2.0",
+    "@angular/upgrade": "18.2.0",
     "@asciidoctor/core": "3.0.4",
     "@asyncapi/web-component": "1.2.26",
     "@danielmoncada/angular-datetime-picker": "17.0.0",
@@ -73,7 +73,7 @@
     "monaco-editor": "0.46.0",
     "ng-file-upload": "12.2.13",
     "ngclipboard": "2.0.0",
-    "ngx-file-helpers": "11.0.0",
+    "ngx-file-helpers": "12.0.0",
     "ngx-markdown": "17.1.1",
     "oidc-client-ts": "2.4.0",
     "prismjs": "1.29.0",
@@ -84,13 +84,13 @@
     "tinycolor2": "1.4.2",
     "unicode": "11.0.1",
     "v-accordion": "1.6.0",
-    "zone.js": "0.14.4"
+    "zone.js": "0.14.10"
   },
   "devDependencies": {
-    "@angular-devkit/architect": "0.1702.1",
-    "@angular-devkit/build-angular": "17.2.1",
-    "@angular/cli": "17.2.1",
-    "@angular/compiler-cli": "17.2.3",
+    "@angular-devkit/architect": "0.1802.0",
+    "@angular-devkit/build-angular": "18.2.0",
+    "@angular/cli": "18.2.0",
+    "@angular/compiler-cli": "18.2.0",
     "@gravitee/ui-schematics": "12.22.2",
     "@storybook/addon-essentials": "7.6.17",
     "@storybook/angular": "7.6.17",
@@ -124,7 +124,7 @@
     "license-check-and-add": "4.0.5",
     "prettier": "3.2.5",
     "storybook": "7.6.17",
-    "typescript": "5.3.3"
+    "typescript": "5.5.4"
   },
   "overrides": {
     "swagger-ui": {
